[Bronze III] Title: 행렬 덧셈, Time: 0 ms, Memory: 1228 KB -BaekjoonHub

+++ 백준/Bronze/2738. 행렬 덧셈/행렬 덧셈.c
@@ -3,7 +3,7 @@
 typedef struct matrix {
     int n_row; // number of rows
     int n_col; // number of columns
-    short data[100][100]; // matrix data
+    int data[100][100];
 } matrix;
 
 matrix input_matrix(int n, int m)
